+++ composeApp/src/iosMain/kotlin/com/xiehe/spine/Platform.ios.kt
@@ -7,3 +7,5 @@ class IOSPlatform : Platform {
 }
 
 actual fun getPlatform(): Platform = IOSPlatform()
+
+actual fun currentHour24(): Int = 12